requirements.txt
protobuf==4.25.2
python-dotenv==1.0.0
streamlit==1.30.0
google-generativeai==0.3.2